@@ -38,7 +38,6 @@
     "@types/otp-generator": "^4.0.2",
     "@typescript-eslint/eslint-plugin": "^7.13.0",
     "@typescript-eslint/parser": "^7.13.0",
-    "@vitejs/plugin-react": "^4.2.1",
     "@vitest/ui": "^1.6.0",
     "autoprefixer": "^10.4.18",
     "chalk": "^5.3.0",
@@ -52,7 +51,7 @@
     "prettier": "^3.1.1",
     "prettier-plugin-svelte": "^3.1.2",
     "prisma": "latest",
-    "svelte": "^4.2.7",
+    "svelte": "^5.0.0-next.166",
     "svelte-check": "^3.6.0",
     "sveltekit-rate-limiter": "^0.5.1",
     "sveltekit-superforms": "^2.12.5",
@@ -65,7 +64,6 @@
   "type": "module",
   "dependencies": {
     "@epic-web/cachified": "^5.2.0",
-    "@epic-web/remember": "^1.0.2",
     "@epic-web/totp": "^1.1.2",
     "@faker-js/faker": "^8.4.1",
     "@paralleldrive/cuid2": "^2.2.2",
