@@ -3,7 +3,7 @@ import './db-setup.ts'
 import '#app/utils/env.server.ts'
 // we need these to be imported first 👆
 import '@testing-library/jest-dom/vitest'
-import { cleanup } from '@testing-library/svelte'
+import { cleanup } from '@testing-library/svelte/svelte5'
 import { afterEach, beforeEach, vi, type MockInstance } from 'vitest'
 import { server } from '../../tests/mocks/index.ts'
 import './custom-matchers.ts'
